Created separate folder for app layer (main.c/h placeholder)
