+++ platformio.ini
@@ -33,6 +33,7 @@ build_flags =
   -I Middlewares/ST/STM32_USB_Device_Library/Class/CDC/Inc
   -I Middlewares/ST/STM32_USB_Device_Library/Core/Inc
   -I lib/
+  -I src/app
   -I src/mcu
   -I src/hw
   -I src/svc
