+++ src/app/main.h
@@ -1,6 +1,6 @@
 #ifndef MAIN_H
 #define MAIN_H
 
-#define MAIN_LOOP_DT_MS 10 // 100 Hz
+#define MAIN_LOOP_DT_MS 20 // 50 Hz
 
 #endif
